Split types file into separate files
